allow changing effect position
