+++ src/ui_plugin/systems/button_handlers.rs
@@ -2,7 +2,6 @@
 use std::collections::HashMap;
 use std::{collections::VecDeque, time::Duration};
 
-use bevy::render::view::RenderLayers;
 use bevy::sprite::collide_aabb::collide;
 use bevy::{prelude::*, window::PrimaryWindow};
 
@@ -19,12 +18,11 @@ use crate::{AddRect, JsonNode, JsonNodeText, NodeType, UiState};
 
 use super::ui_helpers::{
     spawn_modal, ButtonAction, ChangeColor, ChangeTheme, DeleteDoc, DocListItemButton, DrawPencil,
-    Drawing, GenericButton, NewDoc, ParticlesEffect, RawText, SaveDoc, TextPosMode, Tooltip,
-    VeloNode, VeloShape,
+    Drawing, GenericButton, NewDoc, RawText, SaveDoc, TextPosMode, Tooltip, VeloNode, VeloShape,
 };
 use super::{ExportToFile, ImportFromFile, ImportFromUrl, MainPanel, ShareDoc};
 use crate::canvas::arrow::components::{ArrowMeta, ArrowMode};
-use crate::components::{Doc, EffectsCamera, Tab};
+use crate::components::{Doc, Tab};
 use crate::resources::{AppState, FontSystemState, LoadDocRequest, SaveDocRequest};
 use crate::utils::{
     bevy_color_to_cosmic, get_timestamp, load_doc_to_memory, ReflectableUuid, UserPreferences,
@@ -735,107 +733,6 @@ pub fn enable_drawing_mode(
                             text.sections[0].style.color = text.sections[0].style.color.with_a(0.5)
                         }
                     }
-                }
-            }
-            Interaction::Hovered => {}
-            Interaction::None => {}
-        }
-    }
-}
-
-#[cfg(not(target_arch = "wasm32"))]
-pub fn particles_effect(
-    mut query: Query<(&Interaction, &Children), (Changed<Interaction>, With<ParticlesEffect>)>,
-    mut text_style_query: Query<&mut Text, With<ParticlesEffect>>,
-    mut commands: Commands,
-    mut effects: ResMut<Assets<bevy_hanabi::EffectAsset>>,
-    mut effects_camera: Query<&mut Camera, With<EffectsCamera>>,
-    mut effects_query: Query<(&Name, Entity)>,
-) {
-    use bevy_hanabi::prelude::*;
-    use rand::Rng;
-
-    for (interaction, children) in &mut query.iter_mut() {
-        match *interaction {
-            Interaction::Pressed => {
-                for child in children.iter() {
-                    if let Ok(mut text) = text_style_query.get_mut(*child) {
-                        if effects_camera.single_mut().is_active {
-                            text.sections[0].style.color = text.sections[0].style.color.with_a(0.5)
-                        } else {
-                            text.sections[0].style.color = text.sections[0].style.color.with_a(1.)
-                        }
-                    }
-                }
-                if effects_camera.single_mut().is_active {
-                    effects_camera.single_mut().is_active = false;
-                    for (name, entity) in effects_query.iter_mut() {
-                        if name.as_str() == "effect:2d" {
-                            commands.entity(entity).despawn_recursive();
-                        }
-                    }
-                } else {
-                    effects_camera.single_mut().is_active = true;
-                    let mut gradient = Gradient::new();
-                    let mut rng = rand::thread_rng();
-                    gradient.add_key(
-                        0.0,
-                        Vec4::new(
-                            rng.gen_range(0.0..1.0),
-                            rng.gen_range(0.0..1.0),
-                            rng.gen_range(0.0..1.0),
-                            1.0,
-                        ),
-                    );
-                    gradient.add_key(
-                        1.0,
-                        Vec4::new(
-                            rng.gen_range(0.0..1.0),
-                            rng.gen_range(0.0..1.0),
-                            rng.gen_range(0.0..1.0),
-                            0.0,
-                        ),
-                    );
-
-                    let mut size_gradient1 = Gradient::new();
-                    size_gradient1.add_key(0.0, Vec2::splat(0.008));
-                    size_gradient1.add_key(0.3, Vec2::splat(0.012));
-                    size_gradient1.add_key(1.0, Vec2::splat(0.0));
-                    let writer = ExprWriter::new();
-                    let lifetime = writer.lit(5.).uniform(writer.lit(10.)).expr();
-                    let spawner = Spawner::rate(rng.gen_range(10.0..300.0).into());
-                    let effect = effects.add(
-                        EffectAsset::new(32768, spawner, writer.finish())
-                            .with_name("Effect")
-                            .init(InitPositionCircleModifier {
-                                center: Vec3::ZERO,
-                                axis: Vec3::Z,
-                                radius: 0.0001,
-                                dimension: ShapeDimension::Surface,
-                            })
-                            .init(InitVelocityCircleModifier {
-                                center: Vec3::ZERO,
-                                axis: Vec3::Z,
-                                speed: CpuValue::Uniform((0.05, 0.1)),
-                            })
-                            .init(InitAttributeModifier {
-                                attribute: Attribute::LIFETIME,
-                                value: lifetime,
-                            })
-                            .render(SizeOverLifetimeModifier {
-                                gradient: size_gradient1,
-                                screen_space_size: false,
-                            })
-                            .render(ColorOverLifetimeModifier { gradient }),
-                    );
-
-                    commands
-                        .spawn(ParticleEffectBundle {
-                            effect: ParticleEffect::new(effect),
-                            ..default()
-                        })
-                        .insert(Name::new("effect:2d"))
-                        .insert(RenderLayers::layer(2));
                 }
             }
             Interaction::Hovered => {}
